Split skit Runner types into separate files [skip ci]
Extract the six types from Runner.swift into their own files
(HelpersOptions, ToolchainCheckResult, FileOutcome, ProcessResult,
CLIError, SwiftRunOutcome), clearing the one_declaration_per_file
warnings.

Convert two type-returning functions into initializers on their
return type: toolchainCheck -> ToolchainCheckResult.init(libPath:) and
resolveHelpers -> CompiledHelpers.init?(nearInputPath:...). Extract
CompiledHelpers into its own file to host the initializer, which also
clears the pre-existing file_name error on Helpers.swift.

Move toolchainStampFilename (used only by the toolchain check) into
ToolchainCheckResult as a private static, removing a global.
